Back out proactive intelligence — rebuilding separately with its own trigger
Decision (Aditya, June 11): proactive gets its own module + its own
button, sequenced AFTER knowledge-base updates — never inside the
day's-end pipeline. The working scaffold (judge, json-schema decisions,
≤1/day code cap, tier-1 reminders, tier-2 briefings) stays minable in
this branch's history at 67d8078.

Day's-end job is now: editor-idle check → iterative updater → mirror
push → notify, fired by the 'Update Knowledge Base' dev button. Kept:
versioned-summary schema (reminderFlagged/itemDate), Briefings folder
(welcome briefing still writes there), Notify.now. Removed:
Proactive.swift, Store.flaggedSummaries, Notify.schedule/cancel, the
proactive self-test mode.

Verified after removal: build green, incremental 11/11, daysend 5/5
(live Sonnet fold, 7 turns/24s).
