+++ Sentient OS macOS/Documentation/Days-End Job (Living System).md
@@ -1,14 +1,23 @@
 # The Day's-End Job — the Living System (June 11)
 
-Everything that makes Sentient *sentient*, behind ONE entry point: `DaysEndJob.shared.run(store:)`.
-Pipeline: editor-idle check → iterative updater → proactive intelligence → mirror push →
-notification. Idempotent (empty queue = cheap no-op), single-flight (an actor bool; re-triggers
-while running are ignored).
+The knowledge-base lifecycle behind ONE entry point: `DaysEndJob.shared.run(store:)`.
+Pipeline: editor-idle check → iterative updater → mirror push → notification. Idempotent
+(empty queue = cheap no-op), single-flight (an actor bool; re-triggers while running are
+ignored).
 
-**Trigger:** the "Run Proactive Intelligence" dev button in RootView's DEV CONTROLS (it fires
-the whole day's-end pipeline, not just the judge) — deliberately the ONLY trigger until the
-condition-gate scheduler lands. The scheduler will simply call the same function on its own
-clock; no logic lives in the button.
+**Proactive intelligence is deliberately NOT in this pipeline** (June 11 decision): it's
+being built separately with its own trigger, scheduled to run AFTER a knowledge-base update
+(both are agentic jobs over the same vault — sequence, never parallelize). A fully working
+scaffold (judge with `--json-schema` decisions, ≤1/day cap enforced in code, tier-1 scheduled
+reminders, tier-2 agentic briefings with WebSearch) lives in git history at `67d8078` — mine
+it when rebuilding. Two hard-won lessons from its live runs: store date-valued pointers in
+Date's NATIVE `timeIntervalSinceReferenceDate` (epoch conversion round-trips lossily and
+re-judges the newest item), and never call `UNUserNotificationCenter.requestAuthorization`
+from a headless run (it hangs).
+
+**Trigger:** the "Update Knowledge Base" dev button in RootView's DEV CONTROLS — deliberately
+the ONLY trigger until the condition-gate scheduler lands. The scheduler will simply call the
+same function on its own clock; no logic lives in the button.
 
 ## The iterative updater (`VaultUpdater`)
 
@@ -25,6 +34,9 @@ clock; no logic lives in the button.
   **usage limit does NOT** — the half-edited vault is exactly what `--resume` (session id kept
   in memory) continues from. Either way nothing was stamped, so a fresh restart is also correct.
 - **On success:** stamp the sent rows, set `VaultActivity.vaultDirty`.
+- ⚠️ **Sized for daily deltas, not a from-scratch corpus.** After a schema wipe with an
+  existing vault on disk: re-analyze, then stamp the corpus directly (see PR #12's migration
+  note) instead of folding ~everything through Sonnet.
 
 ## Editor-idle + push (`VaultActivity` — the seam, not the feature)
 
@@ -35,41 +47,25 @@ clock; no logic lives in the button.
   run. `DaysEndJob.pushIfDirty()` is the one push path — initial gen calls it too, which
   closes the "nothing auto-pushes yet" gap from Arch §7.
 
-## Proactive Intelligence (`Proactive`)
-
-- **Selection:** reminder-flagged summaries with `createdAt` past the `proactive` cursor.
-- **The judge:** one Sonnet call, NO tools, `--json-schema` →
-  `{"decisions":[{kind: remind|brief|skip, time, text, reason}]}`. The taste law is in the
-  prompt (*most days: nothing; at most ONE per day; only time-sensitive, personally
-  actionable*) **and enforced in code** — first non-skip decision wins, rest dropped.
-- **Tier 1 (remind):** `Notify.schedule(at:)` — past/unparseable time fires now (better a late
-  nudge than a silent drop).
-- **Tier 2 (brief):** a second agentic call — `Read,Glob,Grep,WebSearch,Write`, cwd = vault
-  (context), `--add-dir` the **Briefings folder** (`~/Library/Application
-  Support/SentientOS/Briefings/`, deliberately OUTSIDE the vault so briefings never ride the
-  mirror push) → one `<yyyy-MM-dd> — <slug>.md` + a notification. **Never auto-sends** — drafts
-  and offers only.
-- Failures never block the updater/push; the pointer doesn't advance and tomorrow re-judges.
-
 ## Notifications (`Notify`)
 
-Small `UNUserNotificationCenter` wrapper: `now`/`schedule(at:)→id`/`cancel(id:)`. Permission is
-requested lazily on first use (the real ask moves into onboarding). Quiet by design — no-op
-runs never notify.
+Small `UNUserNotificationCenter` wrapper (`now(title:body:)` only, for now). Permission is
+requested lazily on first use (the real ask moves into onboarding); suppressed entirely under
+`SENTIENT_SELFTEST`. Quiet by design — no-op runs never notify.
 
 ## The welcome briefing
 
 Initial generation's second act (`VaultGenerator.writeWelcomeBriefing()`): a cheap Sonnet pass
 over the freshly built vault that writes "What I learned about you" (portrait + 3–5
-cross-domain connections + what happens next) into Briefings — For You's day-one artifact.
-Best-effort, off the UI path, fired alongside the post-gen mirror push.
+cross-domain connections + what happens next) into the **Briefings folder**
+(`~/Library/Application Support/SentientOS/Briefings/` — deliberately OUTSIDE the vault, so
+For You artifacts never ride the mirror push). Best-effort, off the UI path, fired alongside
+the post-gen mirror push.
 
-## Self-tests (real Sonnet calls — they spend a little budget)
+## Self-test (a real Sonnet call — it spends a little budget)
 
 - `SENTIENT_SELFTEST=daysend SENTIENT_VAULT_ROOT=/tmp/scratch` — fixture vault + seeded queue
   → run → asserts the fold, exact-row stamping, vault change, and the second-run no-op.
   REQUIRES the vault-root override; refuses to run if the mirror is enabled without
   `SENTIENT_MIRROR_BASE` (the push step would clobber the real hosted mirror).
-- `SENTIENT_SELFTEST=proactive` — seeds one actionable + one stale flagged summary → proves
-  judging, the ≤1/day cap, and the pointer (re-run judges nothing).
 - `SENTIENT_VAULT_ROOT` is honored by everything vault-shaped (generator, updater, mirror zip).
