+++ Sentient OS macOS/Notify.swift
@@ -3,11 +3,10 @@
 //  Sentient OS macOS
 //
 //  One small wrapper over UNUserNotificationCenter (Part II §D). Quiet by design — the
-//  callers (day's-end job, proactive reminders) notify only when something actually happened;
-//  no-op runs never make noise. Permission is requested lazily on first use for now; the
-//  real ask moves into onboarding's notification step.
-//
-//  Key methods: now(title:body:) · schedule(at:title:body:) -> id · cancel(id:).
+//  day's-end job notifies only when the vault actually changed; no-op runs never make noise.
+//  Permission is requested lazily on first use for now; the real ask moves into onboarding's
+//  notification step. (Scheduled reminders return with proactive intelligence — git 67d8078
+//  has the schedule/cancel implementation.)
 //
 
 import Foundation
@@ -40,29 +39,5 @@ enum Notify {
         content.sound = .default
         try? await UNUserNotificationCenter.current().add(
             UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: nil))
-    }
-
-    /// Schedule a notification for a future date; returns the request id (for cancel).
-    /// A past date fires immediately — better a late nudge than a silent drop.
-    @discardableResult
-    static func schedule(at date: Date, title: String, body: String) async -> String {
-        let id = UUID().uuidString
-        guard !suppressed else { return id }
-        await requestPermissionIfNeeded()
-        let content = UNMutableNotificationContent()
-        content.title = title
-        content.body = body
-        content.sound = .default
-        let interval = date.timeIntervalSinceNow
-        let trigger = interval > 1
-            ? UNTimeIntervalNotificationTrigger(timeInterval: interval, repeats: false)
-            : nil   // past-dated → deliver now
-        try? await UNUserNotificationCenter.current().add(
-            UNNotificationRequest(identifier: id, content: content, trigger: trigger))
-        return id
-    }
-
-    static func cancel(id: String) {
-        UNUserNotificationCenter.current().removePendingNotificationRequests(withIdentifiers: [id])
     }
 }
